+++ lib/react_webpack_rails/services/camelize_keys.rb
@@ -4,7 +4,7 @@ module ReactWebpackRails
       def self.call(props)
         return props unless props.is_a?(Hash)
         props.inject({}) do |h, (k, v)|
-          h[k.to_s.camelize(:lower)] = v.is_a?(Hash) ? camelize_props_key(v) : v
+          h[k.to_s.camelize(:lower)] = v.is_a?(Hash) ? self.call(v) : v
           h
         end
       end